# airborne_docs/openapi/airborne.openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "Airborne",
        "version": "",
        "description": "Service for managing OTA updates and configurations"
    },
    "paths": {
        "/api/file": {
            "post": {
                "description": "Register a file by URL, recording its metadata and assigning it a version. Pass the organisation and application in the x-organisation and x-application headers. Use this when the file is already hosted somewhere the server can reach; to upload the bytes directly, use UploadFile instead. Requires a bearer token.",
                "operationId": "CreateFile",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateFileRequestContent"
                            }
                        }
                    },
                    "required": true
                },
                "parameters": [
                    {
                        "name": "x-application",
                        "in": "header",
                        "description": "Name of the application",
                        "schema": {
                            "type": "string",
                            "description": "Name of the application"
                        },
                        "required": true
                    },
                    {
                        "name": "x-organisation",
                        "in": "header",
                        "description": "Name of the organisation",
                        "schema": {
                            "type": "string",
                            "description": "Name of the organisation"
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CreateFile 200 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreateFileResponseContent"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "BadRequestError 400 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BadRequestErrorResponseContent"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized 401 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedResponseContent"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "ForbiddenError 403 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenErrorResponseContent"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NotFoundError 404 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundErrorResponseContent"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "InternalServerError 500 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InternalServerErrorResponseContent"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Files"
                ]
            }
        },
...
        "/api/file-groups/{name}": {
            "get": {
                "description": "Get a file group and its full version history by name. Pass the organisation and application in the x-organisation and x-application headers. Requires a bearer token.",
                "operationId": "GetFileGroup",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Name of the file group",
                        "schema": {
                            "type": "string",
                            "description": "Name of the file group"
                        },
                        "required": true
                    },
                    {
                        "name": "x-application",
                        "in": "header",
                        "description": "Name of the application",
                        "schema": {
                            "type": "string",
                            "description": "Name of the application"
                        },
                        "required": true
                    },
                    {
                        "name": "x-organisation",
                        "in": "header",
                        "description": "Name of the organisation",
                        "schema": {
                            "type": "string",
                            "description": "Name of the organisation"
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "GetFileGroup 200 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetFileGroupResponseContent"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "BadRequestError 400 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BadRequestErrorResponseContent"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized 401 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedResponseContent"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "ForbiddenError 403 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenErrorResponseContent"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NotFoundError 404 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundErrorResponseContent"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "InternalServerError 500 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InternalServerErrorResponseContent"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Files"
                ]
            }
        },
        "/api/file-groups/{name}/versions": {
            "post": {
                "description": "Create a new immutable version of a file group, snapshotting the given files with its own metadata. The version number is assigned automatically. Pass the organisation and application in the x-organisation and x-application headers. Requires a bearer token.",
                "operationId": "CreateFileGroupVersion",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateFileGroupVersionRequestContent"
                            }
                        }
                    },
                    "required": true
                },
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Name of the file group",
                        "schema": {
                            "type": "string",
                            "description": "Name of the file group"
                        },
                        "required": true
                    },
                    {
                        "name": "x-application",
                        "in": "header",
                        "description": "Name of the application",
                        "schema": {
                            "type": "string",
                            "description": "Name of the application"
                        },
                        "required": true
                    },
                    {
                        "name": "x-organisation",
                        "in": "header",
                        "description": "Name of the organisation",
                        "schema": {
                            "type": "string",
                            "description": "Name of the organisation"
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CreateFileGroupVersion 200 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreateFileGroupVersionResponseContent"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "BadRequestError 400 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BadRequestErrorResponseContent"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized 401 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedResponseContent"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "ForbiddenError 403 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenErrorResponseContent"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NotFoundError 404 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundErrorResponseContent"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "InternalServerError 500 response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InternalServerErrorResponseContent"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Files"
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "BadRequestErrorResponseContent": {
                "type": "object",
                "description": "Bad request error",
                "properties": {
                    "message": {
                        "type": "string"
                    }
                },
                "required": [
                    "message"
                ]
            },
...
            "CreateFileRequestContent": {
                "type": "object",
                "description": "Create file request",
                "properties": {
                    "file_path": {
                        "type": "string",
                        "description": "Path where the file will be stored on sdk"
                    },
                    "url": {
                        "type": "string",
                        "description": "URL from where the file can be downloaded"
                    },
                    "tag": {
                        "type": "string",
                        "description": "Tag to identify the file"
                    },
                    "metadata": {
                        "description": "Metadata associated with the file in Stringified JSON format or a file attachment"
                    }
                },
                "required": [
                    "file_path",
                    "url"
                ]
            },
...
            "CreateFileResponseContent": {
                "type": "object",
                "description": "Create file response",
                "properties": {
                    "id": {
                        "type": "string",
                        "description": "id of the file"
                    },
                    "file_path": {
                        "type": "string",
                        "description": "Path where the file is stored on sdk"
                    },
                    "url": {
                        "type": "string",
                        "description": "URL from where the file can be downloaded"
                    },
                    "version": {
                        "type": "number",
                        "description": "Version of the file"
                    },
                    "tag": {
                        "type": "string",
                        "description": "Tag associated with the file"
                    },
                    "size": {
                        "type": "number",
                        "description": "Size of the file in bytes"
                    },
                    "checksum": {
                        "type": "string",
                        "description": "Checksum of the file"
                    },
                    "metadata": {
                        "description": "Metadata associated with the file"
                    },
                    "status": {
                        "type": "string",
                        "description": "Status of the file"
                    },
                    "created_at": {
                        "type": "string",
                        "description": "Date of creation of the file"
                    }
                },
                "required": [
                    "checksum",
                    "created_at",
                    "file_path",
                    "id",
                    "metadata",
                    "size",
                    "status",
                    "url",
                    "version"
                ]
            },
...
            "UnauthorizedResponseContent": {
                "type": "object",
                "description": "Unauthorized error",
                "properties": {
                    "message": {
                        "type": "string"
                    }
                },
                "required": [
                    "message"
                ]
            },
...
        },
        "securitySchemes": {
            "smithy.api.httpBearerAuth": {
                "type": "http",
                "description": "HTTP Bearer authentication",
                "scheme": "Bearer"
            }
        }
    },
    "security": [
        {
            "smithy.api.httpBearerAuth": []
        }
    ]
}